feat: Add ListedBook component
As the books in both views are list items, this is a presentational component who renders a book as a part of a list. This is helpful because the specs of how a book should render in a list is in just one spot. I think that I'll find myself writing the same code with similar props on the bookshelf component and the search result component so I'm introducing this to get rid of that.