feat: Add List component
This is the main view of the app. After render it will be initiate the promise to fetch data. This component main objective is to render the Bookshelfs with the correct books. The key are in two props the books and the bookshelfDetails, the first is an array with all the books data and the second is a shortcut map who knows about which book id is on which bookshelf. This allows the renderBookshelfs method to perform a filter on books array based on the bookshelfDetails map.