feat: Add Search component
This is the component who encapsulates the search feature of the app. When the component is mounted it will fetch the list data in order to get the current bookshelf detail. When the component is unmount, the parent search history is deleted in order to always start fresh on navigation. It introduces a debouncing on the input using react-debounce-input to avoid performing N request based on the length of the search term.